pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Simple client wrapper (API) that sits between MPESA and BtoBet."
authors = ["clovisphere <clovis.mugaruka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
alembic = "^1.12.1"
httpx = "^0.25.1"
structlog = "^23.2.0"
gunicorn = "^21.2.0"
fastapi-health = "^0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.5.0"

[tool.pyright]
venvPath = "/Users/clov/.local/share/virtualenvs"
venv = "app-J5gpEhWq-py3.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
